pyproject.toml
[project]
name = "seeweb-gpu-kubernetes-setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.26.0",
    "botocore>=1.42.21",
    "numpy>=1.21.0",
    "pandas>=2.3.3",
    "pillow>=9.0.0",
    "python-dotenv>=0.19.0",
    "rich>=10.0.0",
    "torch>=2.0.0",
    "torchaudio>=2.9.1",
    "torchvision>=0.15.0",
]
